+++ changelog.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.15.0] - 2022-11-30
+
+### Added
+
+- Added `Janitor:RemoveNoClean` and `Janitor:RemoveListNoClean`, which allows removal without cleaning.
+- Added a nicer type return for Janitor. No more `typeof(Janitor.new())`!
+
+### Fixed
+
+- Fixed serious issue where `Janitor:RemoveList` would not actually remove anything from the indices reference.
+
 ## [1.14.2] - 2022-11-09
 
 ### Added